requirements.txt
instagrapi==2.1.2
python-dotenv==1.0.1
Pillow>=10.0.0
moviepy==1.0.3
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
itsdangerous==2.2.0
python-multipart==0.0.20
pydantic==2.10.4
bcrypt==4.2.1
email-validator==2.2.0

